requirements.txt
pandas
numpy
beautifulsoup4
loguru
pyaml
pymongo
PyYAML
pydantic